Improve dashboard configuration.
